perf(tiles): store cover ids in sixteen bits
The renderer holds ground cover in an `R16Uint` texture and the assigned
ids reach 0x080c, but tiles were written as 32-bit samples. The upper
half was always zero: 123 MB of it across a generated tree, and a
narrowing loop over every texel of the raster at every startup.

Tiles are now written `Gray16`. Over the shipped extent the materials
product goes from 247.1 MB to 124.1 MB, and elevation is untouched --
heights are `f32` on disk and `R32Float` on the GPU, so they already
matched and a narrower float would be far worse than useless. Sixteen
bits is 1 to 2 m of vertical quantisation at these elevations, which
would terrace every hillside.

`write_material_tile` keeps taking `u32` and narrows inside. Changing
the signature would have rippled through six call sites in
`terrain-process` and their `Vec<u32>` buffers, for a tool whose output
is already built and is not being re-run. Narrowing in one place also
puts the one thing that can go wrong under one check: an id too wide is
now an error naming the id, not a silent truncation. A truncated id
would produce a perfectly valid tile holding some other material, drawn
somewhere in the far field, with nothing to trace it back to.

Both readers take either width, and that is permanent rather than
transitional. `assets/terrain` is 117 GB of 32-bit tiles and is not
going to be rebuilt to suit a format change.

`MaterialId` becomes `u16`, so the reader now produces exactly the bytes
the upload wants and the narrowing loop in `Terrain::load` is gone
rather than moved. The `as u16` casts through the test modules are that
change reaching its construction sites; the ids involved are enum
discriminants that have always fitted.

Verified against both directions of the compatibility claim:

- A tree regenerated with this writer renders byte-identical frames to
  one written before it, from two cameras. Same landscape, same seed,
  only the sample width differs.
- The installed 32-bit generated tree renders byte-identical to what it
  rendered before the change, and the 117 GB survey tree opens and
  draws its water, snow, vegetation and built-up ground correctly.
- `a_material_id_too_wide_to_store_is_refused` is new, and was checked
  by replacing the conversion with a truncating cast.

`a_material_tile_round_trips` now spreads its ids across the whole
sixteen bits instead of counting upwards. Counting upwards is what made
it fail here first -- it ran past 0x10000 and tripped the new check --
and a version that only wrote small ids would not have noticed a writer
dropping the top byte.

+++ src/scene.rs
@@ -787,10 +787,10 @@ mod tests {
     /// sixteen bits of the material word, the rest being the sub-pixel offset.
     const MATERIAL_MASK: u32 = 0xffff;
 
-    const GRASS: MaterialId = MaterialId(Material::Grass.id());
-    const SAND: MaterialId = MaterialId(Material::Sand.id());
-    const LAKE: MaterialId = MaterialId(Material::Lake.id());
-    const ROCK: MaterialId = MaterialId(Material::BareRock.id());
+    const GRASS: MaterialId = MaterialId(Material::Grass.id() as u16);
+    const SAND: MaterialId = MaterialId(Material::Sand.id() as u16);
+    const LAKE: MaterialId = MaterialId(Material::Lake.id() as u16);
+    const ROCK: MaterialId = MaterialId(Material::BareRock.id() as u16);
     /// An id inside the water block that no version of the enum has assigned.
     const UNASSIGNED: MaterialId = MaterialId(0x0109);
 
@@ -1105,7 +1105,7 @@ mod tests {
                 return None;
             }
             let packed = self.materials[(y * SIZE + x) as usize];
-            Some(MaterialId(packed & MATERIAL_MASK))
+            Some(MaterialId((packed & MATERIAL_MASK) as u16))
         }
 
         /// How many pixels of the frame are sky.
@@ -1394,7 +1394,7 @@ mod tests {
             .map(|index| {
                 let (col, row) = (index % RASTER, index / RASTER);
                 let block = (row / BLOCK) * across + col / BLOCK;
-                MaterialId(Material::ALL[block as usize % Material::ALL.len()].id())
+                MaterialId(Material::ALL[block as usize % Material::ALL.len()].id() as u16)
             })
             .collect();
 
@@ -1437,7 +1437,7 @@ mod tests {
         let uniform = |material: Material| {
             render(
                 vec![0.0; (RASTER * RASTER) as usize],
-                vec![MaterialId(material.id()); (RASTER * RASTER) as usize],
+                vec![MaterialId(material.id() as u16); (RASTER * RASTER) as usize],
                 over_the_checkerboard,
             )
         };
@@ -1474,7 +1474,7 @@ mod tests {
                 let gain = (shown(first.pixel(x, y)) - shown(second.pixel(x, y)))
                     / (first_albedo - second_albedo);
                 let haze = shown(first.pixel(x, y)) - first_albedo * gain;
-                let want = crate::sky::tonemap(albedo(id.0) * gain + haze)
+                let want = crate::sky::tonemap(albedo(u32::from(id.0)) * gain + haze)
                     .to_array()
                     .map(terrain_tiles::linear_to_srgb);
                 let got = frame.pixel(x, y);
@@ -2976,7 +2976,7 @@ mod tests {
         // Every material in the book, tiled: any misregistered window shows
         // as one id where another belongs, and ids compare exactly.
         let materials: Vec<MaterialId> = (0..RASTER * RASTER)
-            .map(|i| MaterialId(Material::ALL[i as usize % Material::ALL.len()].id()))
+            .map(|i| MaterialId(Material::ALL[i as usize % Material::ALL.len()].id() as u16))
             .collect();
 
         let aim = |camera: &mut Camera| {
